Added update to increment the retry count for abandoned messages

+++ HyperMsg.Broker/Services/MessageService.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using HyperMsg.Broker.Data.Entities;
 using HyperMsg.Broker.Data.Repositories;
 using HyperMsg.Contracts;
@@ -15,14 +16,14 @@ namespace HyperMsg.Broker.Services
             _messageRepository = messageRepository;
         }
 
-        public IEnumerable<Message> Get(string endpoint, string count)
+        public IEnumerable<Message> Get(string endPoint, string count)
         {
             int messageCount;
 
             if (!int.TryParse(count, out messageCount) || messageCount < 1) messageCount = 1;
             if (messageCount > 100) messageCount = 100;
 
-            var entities = _messageRepository.Get(endpoint, messageCount);
+            var entities = _messageRepository.Get(endPoint, messageCount);
 
             var messages = new List<Message>();
 
@@ -55,7 +56,19 @@ namespace HyperMsg.Broker.Services
 
         public void Acknowledge(Acknowledgement acknowlege)
         {
-            _messageRepository.Remove(acknowlege.MessageIds);
+            if (acknowlege.IsAbandoned)
+            {
+                var entities = _messageRepository.Get(acknowlege.MessageIds).ToList();
+                entities.ForEach(me => me.RetryCount++);
+
+                // TODO: If any of the entities have reached there retry count then we need to dead letter them
+
+                _messageRepository.Update(entities.ToArray());
+            }
+            else
+            {
+                _messageRepository.Remove(acknowlege.MessageIds);
+            }
         }
     }
 }
